backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
groq>=0.4.1
google-generativeai>=0.4.1
praw>=7.7.1
python-dotenv>=1.0.1
